Break __memory_resource_adaptor into its own header
This diff pulls `__memory_resource_adaptor` out of `function.hpp` and
into its own header.
